ui: Add tooltip class
There are a few components that use a tooltip with their own
implementation. This adds a tooltip class that can be shared between them.

Allow the tooltip to have a custom label, something like a container
with multiple labels each one with different styles.

In the next commits it will be used by authList too.