pointerA11yTimeout: Set the cursor through ClutterActor API
Set the cursor on the stage. This differs with the "root" cursor
approach in that actors with a non-inherited cursor (e.g. I-beam on
text actors) will stick to that instead of showing the cursor as
set here.

If that's turns out not to be enough, might make sense to move this
piece of a11y feedback to Mutter.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3952>