authServicesSwitchable: Allow resetting expired password
JSON protocol can't inform when a password is expired, so it's needed to
check the info message.

When the password is expired, a resetting process will be started. It will
have multiple requests to insert the current password and insert the new
password. This has to be done using the old flow because the PAM JSON protocol
doesn't support this process yet.