status/powerProfiles: Indicate non-default power mode in top bar
It is quite easy to switch the power profile for some specific
purpose, and then forgetting to change it back and experiencing
unexpected performance behavior later.

To help with that, indicate the active power profile in the top
bar if it diverges from the default 'balanced' one.

Closes: https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/7440
Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3961>