capsLockWarning: Mark as not visible to a11y tree when hidden
We're animating it away, but not for the a11y tree.

This fixes it, so the caps lock warning is not read regardless of the state
of the key.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3981>
(cherry picked from commit ccd1a0b8f52e29ccd2356c5d2ac34e473b174cd0)