layout: Export STARTUP_ANIMATION_TIME
We are using it in overviewControls but the value is not exported.

This caused an issue when the animations are not enabled, since in a
racy way (e.g. only sometimes, depending on the init speed) the value
was undefined, and so we ended up falling back to the default animation
duration time.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3982>
(cherry picked from commit 087b0fda5d64f7a74460ef4a0f923a1235aff50f)