layout: Clarify startup animation code
Once the LayoutManager is initialized, we do the startup animation
that consists of 4 steps:

 1. load the system background
 2. prepare the animation
 3. run the animation
 4. complete initialization post-animation

This isn't immediately obvious right now, as the individual steps
are chained in various ways.

It also doesn't help that it's `_loadBackground()` that starts the
chain, as that hardly implies "startup animation".

Clarify that code by adding an async `_doStartupAnimation()` method
that handles all steps in order.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3988>