workspaceAnimation: Split out BaseWorkspaceGroup
Since we started to show DESKTOP windows as part of the background
during the workspace animation, we need to track both the windows
relevant for the monitor/workspace and DESKTOP windows.

We are about to split out the background handling into its own class.

To prepare for this, split out a BaseWorkspaceGroup class that handles the
basic "create clones of relevant windows and track them" functionality
common to both cases.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3995>