brightnessManager: Save and restore the monitor brightnesses
So far we relied on the backlights to somehow magically remember their
state, and for sysfs based backlights this somewhat worked. The
backlight however is a low-level concept that sometimes diverges from
the higher level "user set the slider to X%" concept that the shell has.

For example, if we are dimming the screen for power saving reasons, this
sysfs backlight might remember the dimmed state.

Then there is the reference white based software backlight, which only
exists inside mutter. Nothing is persisted here.

In the future we will have fused backlights where the sysfs and
reference white are adjusted on demand to create a brightness level with
the right amount of HDR headroom. There the backlight and shell monitor
brightness concepts diverge even more.

Finally, even if a backlight were to persist its own state, we have
multiple color modes (HDR, SDR, ...) where a different brightness value
might be required.

This commit takes care of all of this by simply persisting the
user/shell brightness target of the monitors in their color mode.

We use the vendor/product/serial triple to identify monitors, but for
bad EDIDs we fall back on matching the connector as well.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/4006>