workspace: Interpolate window dimensions when transitioning
Since the maximum scale is WINDOW_PREVIEW_MAXIMUM_SCALE (0.95),
dividing the dimensions with WINDOW_PREVIEW_MAXIMUM_SCALE usually
returns the original dimensions of the windows, which we can use
the interpolate the dimensions and coordinates.

However, these results are inconsistent with larger windows,
as the `width` and `height` variables change during transition,
which will make their sizes grow and shrink non-linearly.