overviewControls: Use ease-in-out-quad when transitioning into overview
Same reason as https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/4000.

The first quarter of transitioning into the overview changes the most
compared to the rest, so it's best to start with a slow curve
at the beginning without compromising too much at the end.
Ease-in-out-quad seems to provide the best compromise.